update code style and ci config

+++ app/Http/Controllers/NotebookController.php
@@ -6,6 +6,7 @@ use App\Notebook;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Input;
 use Ramsey\Uuid\Uuid;
+use Illuminate\Support\Facades\Validator;
 
 class NotebookController extends Controller
 {
@@ -17,6 +18,13 @@ class NotebookController extends Controller
 
     public function add(Request $request)
     {
+        $inputs = $request->all();
+        $validator = Validator::make($inputs, [
+            'name' => 'required',
+        ], config('validation')['notebooks']);
+        if ($validator->fails()) {
+            return $this->makeReturnArray($validator->errors()->messages(), 403, '');
+        }
         $notebook = new Notebook();
         $notebook->uuid = strtoupper(Uuid::uuid1()->getHex());
         $notebook->name = Input::input('name');
@@ -28,14 +36,24 @@ class NotebookController extends Controller
     public function delete(Request $request, $uuid)
     {
         $notebook = Notebook::where('uuid', '=', $uuid)->where('uid', '=', $request->user()->id)->first();
+        if (empty($notebook)) {
+            return $this->makeReturnArray([], 404, $this->lang('names.notebooks') . $this->lang('msg.not-found') . ',' . $this->lang('msg.retry'));
+        }
         $notebook->delete();
         return $this->makeReturnArray($notebook);
     }
 
     public function save(Request $request, $uuid)
     {
+        $inputs = $request->all();
+        $validator = Validator::make($inputs, [
+            'name' => 'required',
+        ], config('validation')['users']);
+        if ($validator->fails()) {
+            return $this->makeReturnArray($validator->errors()->messages(), 403, '');
+        }
         $notebook = Notebook::where('uuid', '=', $uuid)->where('uid', '=', $request->user()->id)->first();
-        $notebook->name = Input::input('name');
+        $notebook->name = $request->input('name');
         $notebook->save();
         return $this->makeReturnArray($notebook);
     }
